WIP: compose react part
