+++ client/src/axiom/index.ts
@@ -1,1 +1,2 @@
 export * from './js';
+// NOTE: Do not export './web' because it is only used for the Browser
